adding least square regression maths

+++ src/include/SVD.cpp
@@ -87,7 +87,7 @@ static double PYTHAG(double a, double b)
     return(result);
 }
 
-static int dsvd(float a[][4], int m, int n, float *w, float v[][4])
+static int dsvd(double a[][4], int m, int n, double *w, double v[][4])
 {
     int flag, i, its, j, jj, k, l, nm;
     double c, f, h, s, x, y, z;
@@ -112,69 +112,69 @@ static int dsvd(float a[][4], int m, int n, float *w, float v[][4])
         if (i < m) 
         {
             for (k = i; k < m; k++) 
-                scale += fabs((double)a[k][i]);
+                scale += fabs(a[k][i]);
             if (scale) 
             {
                 for (k = i; k < m; k++) 
                 {
-                    a[k][i] = (float)((double)a[k][i]/scale);
-                    s += ((double)a[k][i] * (double)a[k][i]);
+                    a[k][i] = (a[k][i]/scale);
+                    s += (a[k][i] * a[k][i]);
                 }
-                f = (double)a[i][i];
+                f = a[i][i];
                 g = -SIGN(sqrt(s), f);
                 h = f * g - s;
-                a[i][i] = (float)(f - g);
+                a[i][i] = (f - g);
                 if (i != n - 1) 
                 {
                     for (j = l; j < n; j++) 
                     {
                         for (s = 0.0, k = i; k < m; k++) 
-                            s += ((double)a[k][i] * (double)a[k][j]);
+                            s += (a[k][i] * a[k][j]);
                         f = s / h;
                         for (k = i; k < m; k++) 
-                            a[k][j] += (float)(f * (double)a[k][i]);
+                            a[k][j] += (f * a[k][i]);
                     }
                 }
                 for (k = i; k < m; k++) 
-                    a[k][i] = (float)((double)a[k][i]*scale);
+                    a[k][i] = (a[k][i]*scale);
             }
         }
-        w[i] = (float)(scale * g);
+        w[i] = (scale * g);
     
         /* right-hand reduction */
         g = s = scale = 0.0;
         if (i < m && i != n - 1) 
         {
             for (k = l; k < n; k++) 
-                scale += fabs((double)a[i][k]);
+                scale += fabs(a[i][k]);
             if (scale) 
             {
                 for (k = l; k < n; k++) 
                 {
-                    a[i][k] = (float)((double)a[i][k]/scale);
-                    s += ((double)a[i][k] * (double)a[i][k]);
+                    a[i][k] = (a[i][k]/scale);
+                    s += (a[i][k] * a[i][k]);
                 }
-                f = (double)a[i][l];
+                f = a[i][l];
                 g = -SIGN(sqrt(s), f);
                 h = f * g - s;
-                a[i][l] = (float)(f - g);
+                a[i][l] = (f - g);
                 for (k = l; k < n; k++) 
-                    rv1[k] = (double)a[i][k] / h;
+                    rv1[k] = a[i][k] / h;
                 if (i != m - 1) 
                 {
                     for (j = l; j < m; j++) 
                     {
                         for (s = 0.0, k = l; k < n; k++) 
-                            s += ((double)a[j][k] * (double)a[i][k]);
+                            s += (a[j][k] * a[i][k]);
                         for (k = l; k < n; k++) 
-                            a[j][k] += (float)(s * rv1[k]);
+                            a[j][k] += (s * rv1[k]);
                     }
                 }
                 for (k = l; k < n; k++) 
-                    a[i][k] = (float)((double)a[i][k]*scale);
+                    a[i][k] = (a[i][k]*scale);
             }
         }
-        anorm = MAX(anorm, (fabs((double)w[i]) + fabs(rv1[i])));
+        anorm = MAX(anorm, (fabs(w[i]) + fabs(rv1[i])));
     }
   
     /* accumulate the right-hand transformation */
@@ -185,14 +185,14 @@ static int dsvd(float a[][4], int m, int n, float *w, float v[][4])
             if (g) 
             {
                 for (j = l; j < n; j++)
-                    v[j][i] = (float)(((double)a[i][j] / (double)a[i][l]) / g);
+                    v[j][i] = ((a[i][j] / a[i][l]) / g);
                     /* double division to avoid underflow */
                 for (j = l; j < n; j++) 
                 {
                     for (s = 0.0, k = l; k < n; k++) 
-                        s += ((double)a[i][k] * (double)v[k][j]);
+                        s += (a[i][k] * v[k][j]);
                     for (k = l; k < n; k++) 
-                        v[k][j] += (float)(s * (double)v[k][i]);
+                        v[k][j] += (s * v[k][i]);
                 }
             }
             for (j = l; j < n; j++) 
@@ -207,7 +207,7 @@ static int dsvd(float a[][4], int m, int n, float *w, float v[][4])
     for (i = n - 1; i >= 0; i--) 
     {
         l = i + 1;
-        g = (double)w[i];
+        g = w[i];
         if (i < n - 1) 
             for (j = l; j < n; j++) 
                 a[i][j] = 0.0;
@@ -219,14 +219,14 @@ static int dsvd(float a[][4], int m, int n, float *w, float v[][4])
                 for (j = l; j < n; j++) 
                 {
                     for (s = 0.0, k = l; k < m; k++) 
-                        s += ((double)a[k][i] * (double)a[k][j]);
-                    f = (s / (double)a[i][i]) * g;
+                        s += (a[k][i] * a[k][j]);
+                    f = (s / a[i][i]) * g;
                     for (k = i; k < m; k++) 
-                        a[k][j] += (float)(f * (double)a[k][i]);
+                        a[k][j] += (f * a[k][i]);
                 }
             }
             for (j = i; j < m; j++) 
-                a[j][i] = (float)((double)a[j][i]*g);
+                a[j][i] = (a[j][i]*g);
         }
         else 
         {
@@ -250,7 +250,7 @@ static int dsvd(float a[][4], int m, int n, float *w, float v[][4])
                     flag = 0;
                     break;
                 }
-                if (fabs((double)w[nm]) + anorm == anorm) 
+                if (fabs(w[nm]) + anorm == anorm) 
                     break;
             }
             if (flag) 
@@ -262,28 +262,28 @@ static int dsvd(float a[][4], int m, int n, float *w, float v[][4])
                     f = s * rv1[i];
                     if (fabs(f) + anorm != anorm) 
                     {
-                        g = (double)w[i];
+                        g = w[i];
                         h = PYTHAG(f, g);
-                        w[i] = (float)h; 
+                        w[i] = h; 
                         h = 1.0 / h;
                         c = g * h;
                         s = (- f * h);
                         for (j = 0; j < m; j++) 
                         {
-                            y = (double)a[j][nm];
-                            z = (double)a[j][i];
-                            a[j][nm] = (float)(y * c + z * s);
-                            a[j][i] = (float)(z * c - y * s);
+                            y = a[j][nm];
+                            z = a[j][i];
+                            a[j][nm] = (y * c + z * s);
+                            a[j][i] = (z * c - y * s);
                         }
                     }
                 }
             }
-            z = (double)w[k];
+            z = w[k];
             if (l == k) 
             {                  /* convergence */
                 if (z < 0.0) 
                 {              /* make singular value nonnegative */
-                    w[k] = (float)(-z);
+                    w[k] = (-z);
                     for (j = 0; j < n; j++) 
                         v[j][k] = (-v[j][k]);
                 }
@@ -296,9 +296,9 @@ static int dsvd(float a[][4], int m, int n, float *w, float v[][4])
             }
     
             /* shift from bottom 2 x 2 minor */
-            x = (double)w[l];
+            x = w[l];
             nm = k - 1;
-            y = (double)w[nm];
+            y = w[nm];
             g = rv1[nm];
             h = rv1[k];
             f = ((y - z) * (y + z) + (g - h) * (g + h)) / (2.0 * h * y);
@@ -311,7 +311,7 @@ static int dsvd(float a[][4], int m, int n, float *w, float v[][4])
             {
                 i = j + 1;
                 g = rv1[i];
-                y = (double)w[i];
+                y = w[i];
                 h = s * g;
                 g = c * g;
                 z = PYTHAG(f, h);
@@ -324,13 +324,13 @@ static int dsvd(float a[][4], int m, int n, float *w, float v[][4])
                 y = y * c;
                 for (jj = 0; jj < n; jj++) 
                 {
-                    x = (double)v[jj][j];
-                    z = (double)v[jj][i];
-                    v[jj][j] = (float)(x * c + z * s);
-                    v[jj][i] = (float)(z * c - x * s);
+                    x = v[jj][j];
+                    z = v[jj][i];
+                    v[jj][j] = (x * c + z * s);
+                    v[jj][i] = (z * c - x * s);
                 }
                 z = PYTHAG(f, h);
-                w[j] = (float)z;
+                w[j] = z;
                 if (z) 
                 {
                     z = 1.0 / z;
@@ -341,15 +341,15 @@ static int dsvd(float a[][4], int m, int n, float *w, float v[][4])
                 x = (c * y) - (s * g);
                 for (jj = 0; jj < m; jj++) 
                 {
-                    y = (double)a[jj][j];
-                    z = (double)a[jj][i];
-                    a[jj][j] = (float)(y * c + z * s);
-                    a[jj][i] = (float)(z * c - y * s);
+                    y = a[jj][j];
+                    z = a[jj][i];
+                    a[jj][j] = (y * c + z * s);
+                    a[jj][i] = (z * c - y * s);
                 }
             }
             rv1[l] = 0.0;
             rv1[k] = f;
-            w[k] = (float)x;
+            w[k] = x;
         }
     }
     free((void*) rv1);
